Move class into servicecatalog/client
