+++ pkg/class/describe_cmd.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/Azure/service-catalog-cli/pkg/command"
 	"github.com/Azure/service-catalog-cli/pkg/output"
+	"github.com/Azure/service-catalog-cli/pkg/service-catalog/client"
 	"github.com/Azure/service-catalog-cli/pkg/traverse"
 	"github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
 	"github.com/spf13/cobra"
@@ -61,9 +62,9 @@ func (c *describeCmd) describe(key string) error {
 	var class *v1beta1.ClusterServiceClass
 	var err error
 	if c.lookupByUUID {
-		class, err = retrieveByUUID(c.Client, key)
+		class, err = client.RetrieveClassByID(c.Client, key)
 	} else {
-		class, err = RetrieveByName(c.Client, key)
+		class, err = client.RetrieveClassByName(c.Client, key)
 	}
 	if err != nil {
 		return err
